+++ CHANGELOG.md
@@ -94,6 +94,10 @@
 
 #### Date: 2/22/2026
 
+### :rocket: Features
+
+-   Added `analytics.recordEvent()` to also send events to PostHog when configured and enabled, while continuing to send to Simple Analytics when available, and returning an error when neither analytics provider is available.
+
 ### :bug: Bug Fixes
 
 -   Fixed freeze caused by pressing <kbd>Shift</kbd> key on keyboard form.